pyproject.toml
[tool.poetry]
name = "claude-proxy"
version = "0.1.0"
description = "Corporate proxy server for Claude Code"
authors = ["Admin"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.32"}
uvloop = "^0.21"
httptools = "^0.6"
httpx = {extras = ["socks"], version = "^0.28"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.30"
alembic = "^1.14"
pydantic-settings = "^2.7"
loguru = "^0.7"
python-jose = {extras = ["cryptography"], version = "^3.3"}
bcrypt = "^5.0"
cryptography = "^44"
jinja2 = "^3.1"
python-multipart = "^0.0.18"
email-validator = "^2.0"
apscheduler = "^3.10"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-asyncio = "^0.24"
anyio = {extras = ["trio"], version = "^4.0"}
httpx = "^0.28"
pytest-cov = "^5.0"
ruff = "^0.15.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
